Miscellaneous improvements
Added a new way to set the editor mode.
Added a way to get a mode from the level editor.
Added a safety check when adding resources.
Added a way to remove resources.
Added event for creating object from save data.
Added a toggle to include inspector in inspector.
Made a lot of fields optional in the inspector UI.