A lot of misc changes.
Added LoadLevelFromBytes/Json to save manager.
Changed ILevelEditorResources to return a IReadOnlyList instead of an array with GetResources.
Changed customData to be public so you can interact with it in the save data.
Fixed structs with single character field names couldn't be generated with ALE compatibility 0.
Fixed field types not being added to the resolver and/or generated.
Fixed issues with cross-assembly references for types.
Fixed loading a level erroring out if objects couldn't be found. Now they just logs errors but doesn't stop the loading.
Fixed compiler issues if you weren't using localization.